requirements.txt
hidapi~=0.11.0.post2
loguru~=0.5.3
schedule~=1.1.0
prometheus-client~=0.12.0
pymongo~=3.12.1
pydantic~=1.8.2
